исправлены мелкие ошибки

+++ game.py
@@ -93,15 +93,18 @@ class Game:
                 kill_pig_collision[0].hp -= 3
                 self.text7 = f2.render(str(kill_pig_collision[0].hp), True, WHITE)
                 if kill_pig_collision[0].hp <= 0:
+                    kill_pig_collision[0].hp = 0
                     self.text7 = f2.render(str(kill_pig_collision[0].hp), True, WHITE)
                     pig.updat = 0
 
             if len(pig_eat_collision) >= 1:
                 if pig.updat == 0:
-                    player.h_p += 2
-                    self.text2 = f1.render(str(player.h_p), True, WHITE)
-                    pig.updat = 1
-                    pig.kill()
+                    if player.h_p + 2 <= 20:
+                        player.h_p += 2
+                        self.text7 = f2.render('', True, WHITE)
+                        self.text2 = f1.render(str(player.h_p), True, WHITE)
+                        pig.updat = 1
+                        pig.kill()
                 else:
                     player.h_p -= 1
             if len(сartridges_collision) >= 1:
